ISSUES-19 fix integration testing
